refactor: move ExtensionManifest data model into homeboy-extension-contract
The keystone Stage-2 cut: ExtensionManifest itself moves to the contract crate.

After all its field types landed in contract crates over cuts 1-11 and the
sidecar split (this branch's parent), ExtensionManifest's struct is pure data.
Its 52 pure accessor methods travel with it; two run-dir-dependent sidecar
helpers (structured_sidecars, structured_sidecar_schema_version) become core
free functions (re-exported from the extension module); semver() and
validate_notification_transports() move too since parse_extension_version,
NotificationTransportConfig::validate, and the required homeboy_error
constructors are all reachable from the leaf crate.

The impl ConfigEntity for ExtensionManifest STAYS in core: ConfigEntity is a
core trait (legal orphan impl for the now-foreign type), and it uses core-only
paths::extension_manifest. Core re-exports ExtensionManifest so all
crate::extension::ExtensionManifest paths stay stable.

This puts the entire extension manifest data model in the leaf contract crate -
the milestone that lets code-audit (which needs ExtensionManifest) depend on
the contract instead of core, and sets up the eventual homeboy-extension crate.
