+++ crates/homeboy-core/src/extension/mod.rs
@@ -88,29 +88,29 @@ pub use lifecycle::{
 };
 pub use maintenance::{exec_tool, update_all};
 pub use manifest::{
-    ActionConfig, ActionType, AgentRuntimeManifestConfig, AuditCapability, AutofixVerifyConfig,
-    BehaviorScenarioNames, BenchConfig, BuildConfig, CiCapability, CiJobFidelity, CiJobMapping,
-    CiJobSpec, CiLocalContext, CiProfileSpec, CliAutoFlag, CliAutoFlagCondition, CliConfig,
-    CliHelpConfig, ComponentEnvConfig, DatabaseCliConfig, DatabaseConfig, DeployCapability,
-    DeployOverride, DeployOwnerHint, DeployVerification, DepsConfig, DiscoveryConfig,
-    DiscoveryMarkerConfig, DocTarget, ExecutableCapability, ExtensionContractProducer,
-    ExtensionContractProducerInvocation, ExtensionContractProducerOutput,
-    ExtensionContractProducerOutputKind, ExtensionContractProducerPhase,
-    ExtensionDiagnosticsConfig, ExtensionManifest, ExtensionMaterializationHelperManifestRef,
-    ExtensionMaterializationSourceContract, ExtensionMaterializationSourceKind,
-    ExtensionToolDiagnosticDeclaration, FeatureContextRule, FileContainsCondition, FuzzConfig,
-    HttpMethod, IncludeWrapperPolicy, InputConfig, LintChangedFileRoute, LintConfig,
-    NotificationTransportConfig, PackageNameSource, PlatformCapability, ProvidesConfig,
-    ReleasePreflightConfig, RemotePathInferenceRule, RemotePathRootRule, RequirementsConfig,
-    RuntimeConfig, RuntimeRequirementsConfig, ScriptsConfig, SelectOption, SettingConfig,
-    SinceTagConfig, SourceSnapshotConfig, StructuredSidecarDeclaration,
-    TestChangedFileExclusiveEnv, TestChangedFileRouting, TestChangedFileRoutingStrategy,
-    TestConfig, TestDriftConfig, TestMappingConfig, TestPassthroughFilter,
-    TestPassthroughFilterStrategy, TestVacuityPolicy, TraceBrowserArtifactMapConfig,
-    TraceBrowserEvidenceAdapterConfig, TraceBrowserMetricAliasConfig,
-    TraceBrowserSummaryAliasConfig, TraceConfig, VersionPatternConfig,
-    EXTENSION_CONTRACT_PRODUCER_SCHEMA, EXTENSION_MATERIALIZATION_SOURCE_SCHEMA,
-    NOTIFICATION_TRANSPORT_SCHEMA,
+    structured_sidecar_schema_version, structured_sidecars, ActionConfig, ActionType,
+    AgentRuntimeManifestConfig, AuditCapability, AutofixVerifyConfig, BehaviorScenarioNames,
+    BenchConfig, BuildConfig, CiCapability, CiJobFidelity, CiJobMapping, CiJobSpec, CiLocalContext,
+    CiProfileSpec, CliAutoFlag, CliAutoFlagCondition, CliConfig, CliHelpConfig, ComponentEnvConfig,
+    DatabaseCliConfig, DatabaseConfig, DeployCapability, DeployOverride, DeployOwnerHint,
+    DeployVerification, DepsConfig, DiscoveryConfig, DiscoveryMarkerConfig, DocTarget,
+    ExecutableCapability, ExtensionContractProducer, ExtensionContractProducerInvocation,
+    ExtensionContractProducerOutput, ExtensionContractProducerOutputKind,
+    ExtensionContractProducerPhase, ExtensionDiagnosticsConfig, ExtensionManifest,
+    ExtensionMaterializationHelperManifestRef, ExtensionMaterializationSourceContract,
+    ExtensionMaterializationSourceKind, ExtensionToolDiagnosticDeclaration, FeatureContextRule,
+    FileContainsCondition, FuzzConfig, HttpMethod, IncludeWrapperPolicy, InputConfig,
+    LintChangedFileRoute, LintConfig, NotificationTransportConfig, PackageNameSource,
+    PlatformCapability, ProvidesConfig, ReleasePreflightConfig, RemotePathInferenceRule,
+    RemotePathRootRule, RequirementsConfig, RuntimeConfig, RuntimeRequirementsConfig,
+    ScriptsConfig, SelectOption, SettingConfig, SinceTagConfig, SourceSnapshotConfig,
+    StructuredSidecarDeclaration, TestChangedFileExclusiveEnv, TestChangedFileRouting,
+    TestChangedFileRoutingStrategy, TestConfig, TestDriftConfig, TestMappingConfig,
+    TestPassthroughFilter, TestPassthroughFilterStrategy, TestVacuityPolicy,
+    TraceBrowserArtifactMapConfig, TraceBrowserEvidenceAdapterConfig,
+    TraceBrowserMetricAliasConfig, TraceBrowserSummaryAliasConfig, TraceConfig,
+    VersionPatternConfig, EXTENSION_CONTRACT_PRODUCER_SCHEMA,
+    EXTENSION_MATERIALIZATION_SOURCE_SCHEMA, NOTIFICATION_TRANSPORT_SCHEMA,
 };
 pub use refactor_protocol::{
     run_refactor_script, run_refactor_script_result, AdjustedItem, ParsedItem,
