+++ crates/homeboy-extension-contract/src/lib.rs
@@ -16,6 +16,7 @@ pub mod core_compat;
 pub mod exec_context;
 pub mod extension_contract_producer;
 pub mod fuzz_config;
+pub mod manifest;
 pub mod manifest_action_config;
 pub mod manifest_capabilities;
 pub mod manifest_capability_config;
@@ -25,6 +26,7 @@ pub mod manifest_toolchain_config;
 pub mod notification_transport_config;
 pub mod runner_contract;
 pub mod sidecar_config;
+pub use manifest::ExtensionManifest;
 pub mod source_metadata_repair;
 pub mod test_drift;
 pub mod trace_config;
